[106] updated file structure for Eco Friends API
